@@ -4,13 +4,13 @@ import 'package:flutterfire_ui/auth.dart';
 
 class AuthGate extends StatelessWidget {
   final Widget app;
-  const AuthGate({Key key, this.app}) : super(key: key);
+  const AuthGate({Key? key, required this.app}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return StreamBuilder(
       stream: FirebaseAuth.instance.authStateChanges(),
-      builder: (BuildContext context, AsyncSnapshot<User> snapshot) {
+      builder: (BuildContext context, AsyncSnapshot<User?> snapshot) {
         if (!snapshot.hasData) {
           return const MaterialApp(
             home: SignInScreen(
