@@ -1,6 +1,6 @@
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
-import 'AuthGate.dart';
+import 'auth_gate.dart';
 import 'app.dart';
 import 'firebase_options.dart';
 
